Implement this temporary file guy to reduce race conditions.
